requirements.txt
pandas
openpyxl
watchdog
python-barcode
Pillow
reportlab
python-escpos
tinydb
pyinstaller
tk
pywin32
